fixup merged class-all-tests.php
